test(survey unit presenter): relocate common logic test
