+++ pearljam-api/src/test/java/fr/insee/pearljam/api/surveyunit/presenter/SurveyUnitAssignedApiPresenterTest.java
@@ -1,11 +1,9 @@
 package fr.insee.pearljam.api.surveyunit.presenter;
 
 import fr.insee.pearljam.domain.surveyunit.model.StateType;
-import fr.insee.pearljam.domain.surveyunit.port.in.SurveyUnitAssignedPresenter;
 import fr.insee.pearljam.domain.surveyunit.readmodel.SurveyUnitAssigned;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.data.domain.Page;
@@ -91,24 +89,6 @@ class SurveyUnitAssignedApiPresenterTest {
     void toStateType_shouldThrowException_whenValueIsInvalid(String value) {
         assertThatThrownBy(() -> presenter.toStateType(value))
             .isInstanceOf(IllegalArgumentException.class);
-    }
-
-
-    @ParameterizedTest
-    @CsvSource(value = {
-        "Jean,Dupont,Jean Dupont",
-        "Marie,Martin,Marie Martin",
-        "Élodie,Dubois,Élodie Dubois",
-        "Jean-Pierre,Durand,Jean-Pierre Durand",
-        "Jean Claude,Martin,Jean Claude Martin",
-        "Louis de,Fontaine,Louis de Fontaine",
-        ",Durand,Durand",
-        "Marie,,Marie",
-        ",, ''"
-    })
-    void buildInterviewerLabel_shouldHandleFrenchNames(String firstName, String lastName, String expected) {
-        assertThat(SurveyUnitAssignedPresenter.buildInterviewerLabel(firstName, lastName))
-            .isEqualTo(expected);
     }
 
 }
